refactor: rename ui/Panel.*
