+++ src/pages/Panel/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 
-import Panel from '../../ui/Panel';
+import Panel from '../../ui/index';
 
 render(<Panel />, window.document.querySelector('#panel'));
